refactoring declaration_sources
